Cleanup code for readability / maintainability
- move variables panel to separate widget
- use templates
