+++ src/airunner/enums.py
@@ -212,6 +212,9 @@ class SignalCode(Enum):
     PLAYBACK_DEVICE_CHANGED = enum.auto()
     RECORDING_DEVICE_CHANGED = enum.auto()
     NODE_EXECUTION_COMPLETED_SIGNAL = enum.auto()
+    CLEAR_WORKFLOW_SIGNAL = enum.auto()
+    WORKFLOW_LOAD_SIGNAL = enum.auto()
+    REGISTER_GRAPH_SIGNAL = enum.auto()
 
 
 class EngineResponseCode(Enum):
