+++ src/airunner/gui/widgets/nodegraph/templates/node_graph.ui
@@ -201,7 +201,7 @@
      <property name="orientation">
       <enum>Qt::Orientation::Horizontal</enum>
      </property>
-     <widget class="QWidget" name="variables" native="true">
+     <widget class="VariablesPanelWidget" name="variables" native="true">
       <layout class="QGridLayout" name="gridLayout_4">
        <property name="leftMargin">
         <number>0</number>
@@ -253,6 +253,14 @@
    </item>
   </layout>
  </widget>
+ <customwidgets>
+  <customwidget>
+   <class>VariablesPanelWidget</class>
+   <extends>QWidget</extends>
+   <header>airunner/gui/widgets/nodegraph/variables_panel</header>
+   <container>1</container>
+  </customwidget>
+ </customwidgets>
  <resources>
   <include location="../../../resources/feather.qrc"/>
  </resources>
