+++ src/airunner/gui/widgets/nodegraph/templates/variables_panel_ui.py
@@ -1,0 +1,51 @@
+# -*- coding: utf-8 -*-
+
+################################################################################
+## Form generated from reading UI file 'variables_panel.ui'
+##
+## Created by: Qt User Interface Compiler version 6.7.0
+##
+## WARNING! All changes made in this file will be lost when recompiling UI file!
+################################################################################
+
+from PySide6.QtCore import (QCoreApplication, QDate, QDateTime, QLocale,
+    QMetaObject, QObject, QPoint, QRect,
+    QSize, QTime, QUrl, Qt)
+from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
+    QFont, QFontDatabase, QGradient, QIcon,
+    QImage, QKeySequence, QLinearGradient, QPainter,
+    QPalette, QPixmap, QRadialGradient, QTransform)
+from PySide6.QtWidgets import (QApplication, QGridLayout, QListWidget, QListWidgetItem,
+    QPushButton, QSizePolicy, QWidget)
+
+class Ui_variables_panel(object):
+    def setupUi(self, variables_panel):
+        if not variables_panel.objectName():
+            variables_panel.setObjectName(u"variables_panel")
+        variables_panel.resize(400, 300)
+        self.gridLayout = QGridLayout(variables_panel)
+        self.gridLayout.setObjectName(u"gridLayout")
+        self.gridLayout.setContentsMargins(0, 0, 0, 0)
+        self.variables_list_widget = QListWidget(variables_panel)
+        self.variables_list_widget.setObjectName(u"variables_list_widget")
+        self.variables_list_widget.setDragEnabled(True)
+
+        self.gridLayout.addWidget(self.variables_list_widget, 0, 0, 1, 1)
+
+        self.pushButton = QPushButton(variables_panel)
+        self.pushButton.setObjectName(u"pushButton")
+
+        self.gridLayout.addWidget(self.pushButton, 1, 0, 1, 1)
+
+
+        self.retranslateUi(variables_panel)
+        self.pushButton.clicked.connect(variables_panel.on_add_variable)
+
+        QMetaObject.connectSlotsByName(variables_panel)
+    # setupUi
+
+    def retranslateUi(self, variables_panel):
+        variables_panel.setWindowTitle(QCoreApplication.translate("variables_panel", u"Form", None))
+        self.pushButton.setText(QCoreApplication.translate("variables_panel", u"Add Variable", None))
+    # retranslateUi
+
